Make a proper Python wrapper over the library bindings
